+++ component/exporter.go
@@ -68,20 +68,29 @@ type ExporterFactory interface {
 	// tests of any implementation of the Factory interface.
 	CreateDefaultConfig() config.Exporter
 
-	// CreateTracesExporter creates a trace exporter based on this config.
+	// CreateTracesExporter creates a TracesExporter based on this config.
 	// If the exporter type does not support tracing or if the config is not valid,
 	// an error will be returned instead.
 	CreateTracesExporter(ctx context.Context, set ExporterCreateSettings, cfg config.Exporter) (TracesExporter, error)
 
-	// CreateMetricsExporter creates a metrics exporter based on this config.
+	// TracesExporterStability gets the stability level of the TracesExporter.
+	TracesExporterStability() StabilityLevel
+
+	// CreateMetricsExporter creates a MetricsExporter based on this config.
 	// If the exporter type does not support metrics or if the config is not valid,
 	// an error will be returned instead.
 	CreateMetricsExporter(ctx context.Context, set ExporterCreateSettings, cfg config.Exporter) (MetricsExporter, error)
 
-	// CreateLogsExporter creates an exporter based on the config.
+	// MetricsExporterStability gets the stability level of the MetricsExporter.
+	MetricsExporterStability() StabilityLevel
+
+	// CreateLogsExporter creates a LogsExporter based on the config.
 	// If the exporter type does not support logs or if the config is not valid,
 	// an error will be returned instead.
 	CreateLogsExporter(ctx context.Context, set ExporterCreateSettings, cfg config.Exporter) (LogsExporter, error)
+
+	// LogsExporterStability gets the stability level of the LogsExporter.
+	LogsExporterStability() StabilityLevel
 }
 
 // ExporterFactoryOption apply changes to ExporterOptions.
@@ -146,6 +155,18 @@ type exporterFactory struct {
 	CreateTracesExporterFunc
 	CreateMetricsExporterFunc
 	CreateLogsExporterFunc
+}
+
+func (e exporterFactory) TracesExporterStability() StabilityLevel {
+	return e.getStabilityLevel(config.TracesDataType)
+}
+
+func (e exporterFactory) MetricsExporterStability() StabilityLevel {
+	return e.getStabilityLevel(config.MetricsDataType)
+}
+
+func (e exporterFactory) LogsExporterStability() StabilityLevel {
+	return e.getStabilityLevel(config.LogsDataType)
 }
 
 // WithTracesExporter overrides the default "error not supported" implementation for CreateTracesExporter and the default "undefined" stability level.